# ApiSpec/swagger.yaml
swagger: '2.0'
info:
  version: v1
  title: REST API Documentation for Schoolbus
  description: API Sample
  termsOfService: None
basePath: /api
paths:
...
  /users:
    get:
      description: Returns all users
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'    
        404:
           description: User not found
    post:
      description: Create new user
      tags:
        - User
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        201:
          description: User created
          headers:
            Location: 
              description: A link to the User
              type: string
              format: url
          schema:
            $ref: "#/definitions/User"
...
  /users/bulk:
    post:
      description: Adds a number of users
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: items
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
  /users/{id}:
    get:
      description: Returns data for a particular user
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'    
        404:
           description: User not found
    put:
      description: Updates a user
      tags:
        - User
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to update
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        404:
          description: User not found
    delete:
      description: Deletes a user
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to delete
          required: true
          type: integer    
      responses:
        200:
          description: OK
        404:
          description: User not found
  /users/{id}/roles:
    get:
      description: Returns the roles for a user
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserRole'    
        404:
           description: User not found
    post:
      description: Adds a role to a user
      tags:
        - User
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to update
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserRole"
      responses:
        201:
          description: Role created for user
          headers:
            Location: 
              description: A link to the User Role
              type: string
              format: url
          schema:
            $ref: "#/definitions/UserRole"
    put:
      description: Updates the roles for a user
      tags:
        - User
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to update
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserRole'
        404:
          description: User not found
  /users/{id}/permissions:
    get:
      description: Returns the set of permissions for a user
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Permission'    
        404:
           description: User not found
  /users/{id}/groups:
    get:
      description: Returns all groups that a user is a member of
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/GroupMembership'    
        404:
           description: User not found
    put:
      description: Updates the active set of groups for a user
      tags:
        - User
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of User to update
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/GroupMembership'
        404:
          description: User not found
  /users/{id}/favourites:
    get:
      description: Returns a user's favourites of a given context type
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserFavourite'
        "404":
           description: User not found
      parameters:
        - name: id
          in: path
          description: id of User to fetch favorites for
          required: true
          type: integer
  /users/{id}/notification:
    get:
      description: Returns a user's notifications
      tags:
        - User
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Notification'
      parameters:
        - name: id
          in: path
          description: id of User to fetch notifications for
          required: true
          type: integer
...
definitions:
  SchoolBus:
    type: object
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      SchoolBusOwner:
        $ref: '#/definitions/SchoolBusOwner'
      IsActive: 
        type: boolean
      IsOutOfProvince:
        type: boolean
      NextInspection:
        type: string
        format: date-time
      CRNO:
        type: string
        description: CR Number.
      NameOfIndependentSchool:
        type: string
      CCWData:
        $ref: '#/definitions/CCWData'
...
  SchoolBusOwner:
    type: object
    required:
      - id    
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key  
        x-type-options:
          min: 1
          max: 1000
      Name:
        type: string
      IsActive:
        type: boolean
      DateCreated:
        type: string
        format: date-time
      PrimaryContact:
        $ref: '#/definitions/SchoolBusOwnerContact'
      LocalArea:
        $ref: '#/definitions/LocalArea'
      City: 
        $ref: '#/definitions/City'
...
  SchoolBusOwnerContact:
    type: object
    required:
      - id    
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      SchoolBusOwner:
          $ref: '#/definitions/SchoolBusOwner'    
...
  CCWData:
    type: object
    required:
      - id    
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      ModelYear:
        format:  int32
        type: integer
        description: Vehicle Year
        x-type-options:
          min: 1970
          max: 2017
      RateClass:
        type: string
      CVIPDecal:
        type: string
      FleetUnitNo: 
        type: integer
      GVW:
        type: integer
      Body:
        type: string
      RebuiltStatus:
        type: string
      CVIPExpiry:
        type: string
        format: date-time
      NetWt:
        type: integer
      Model:
        type: string
      Fuel:
        type: string
      SeatingCapacity:
        type: integer
      Colour:
        type: string
...
  FavouriteContextType:
    type: object
    required:
      - id    
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      Name:
        type: string
        description:  Context Name
        x-chance-type: name
  UserFavourite:
    type: object
    required:
      - id  
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      Name: 
        type: string
        description:  Context Name
        x-chance-type: name
      Value:
        type: string
        description:  Saved search
        x-chance-type: word
      IsDefault:
        type: boolean
      FavouriteContextType:
        $ref: '#/definitions/FavouriteContextType'
  Region:
    type: object
    required:
      - id  
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      Name:
        type: string
        format: string
        description:  Region Name
        x-chance-type: name
  City:
    type: object
    required:
      - id  
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      Name:
        type: string
        format: string
        description:  City Name
        x-chance-type: name
      Region:
        $ref: '#/definitions/Region'
  LocalArea:
    type: object
    required:
      - id  
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      Name:
          type: string
          format: string
          description:  City Name
          x-chance-type: name          
      Region:
        $ref: '#/definitions/Region'
...
  Notification:
    type: object
    required:
      - id
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      Event:
          $ref: '#/definitions/NotificationEvent'
      Event2:
          $ref: '#/definitions/NotificationEvent'
      HasBeenViewed:
        type: boolean
      IsWatchNotification:
        type: boolean
      IsExpired:
        type: boolean
      IsAllDay:
        type: boolean
      PriorityCode: 
        type: string
      User:
        $ref: '#/definitions/User'
  NotificationEvent:
    type: object
    required:
      - id
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      EventTime:
        type: string
        format: datetime
      EventTypeCode:
        type: string
      EventSubTypeCode:
        type: string
      Notes:
        type: string
      NotificationGenerated:
        type: boolean
      SchoolBus:
        $ref: '#/definitions/SchoolBus'
  User:
    type: object
    required:
      - id
      - active
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      givenName:
        type: string
        x-chance-type: name
      surname:
        type: string
        x-chance-type: name
      initials:
        type: string
      email:
        type: string
        x-chance-type: email
      active:
        type: boolean
      smUserId:
        type: string
        description: Security Manager User ID
        x-chance-type: name
      guid:
        type: string
      smAuthorizationDirectory:
        type: string
      userRoles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      groupMemberships:
        type: array
        items:
          $ref: '#/definitions/GroupMembership'
  UserRole:
    type: object
    required:
      - id
      - effectiveDate
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      effectiveDate:
        type: string
        format: date-time
      expiryDate:
        type: string
        format: date-time
      user:
        $ref: '#/definitions/User'
      role:
        $ref: '#/definitions/Role'
  Role:
    type: object
    required:
      - id
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence
      rolePermissions:
        type: array
        items:
          $ref: '#/definitions/RolePermission'
      userRoles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
  RolePermission:
    type: object
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      role:
        $ref: '#/definitions/Role'
      permision:
        $ref: '#/definitions/Permission'
  Permission:
    type: object
    required:
      - id
      - code
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      code:
        type: string
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence
  GroupMembership:
    type: object
    required:
      - id
      - active
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      active:
        type: boolean
      group:
        $ref: '#/definitions/Group'
      user:
        $ref: '#/definitions/User'
  Group:
    type: object
    required:
      - id
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence        
securityDefinitions: {}